requirements.txt
streamlit
gdown==5.2.0
asttokens==2.0.5
Brotli==1.0.9
certifi==2024.7.4
charset-normalizer==3.3.2
cirpy
colorama==0.4.6
comm==0.2.1
ConfigArgParse==1.4
debugpy==1.6.7
decorator==5.1.1
executing==0.8.3
filelock==3.13.1
fsspec==2024.6.1
idna==3.7
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
packaging==24.1
parso==0.8.3
pillow==10.4.0
platformdirs==3.10.0
prompt-toolkit==3.0.43
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.15.1
PySocks==1.7.1
python-dateutil==2.9.0post0
PyYAML==6.0.1
pyzmq==25.1.2
rdkit==2024.3.5
requests==2.32.3
setuptools==72.1.0
six==1.16.0
stack-data==0.2.0
sympy==1.12
torch==2.4.0
torchaudio==2.4.0
torchtext==0.4.0
torchvision==0.19.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.2
wcwidth==0.2.5
wheel==0.43.0
win-inet-pton==1.1.0
transformers
huggingface_hub
